Credentials.DynamoDB.Item

Description

This module contains the schema that is used by Credentials.DynamoDB to serialise encryption parameters to DynamoDB items.

Documentation

padding :: Text

newtype Version

The DynamoDB field used for optimistic locking.

Serialisation of Version handles left-padding to support consistent lexicographic ordering when used as a range in DynamoDB.

Constructors

Version Integer

equals :: Item a => a -> HashMap Text Condition

nameField :: Text

revisionField :: Text

versionField :: Text

wrappedKeyField :: Text

ciphertextField :: Text

digestField :: Text

class Item a where

Methods

toItem :: a -> HashMap Text AttributeValue

Encode an item as a set of attributes including their schema.

parseItem :: HashMap Text AttributeValue -> Either CredentialError a

Decode an item from a set of attributes.

fromItem :: (MonadThrow m, Item a) => HashMap Text AttributeValue -> m a

Decode an item by throwing a CredentialError exception when an error is encountered.

parse :: Attribute a => Text -> HashMap Text AttributeValue -> Either CredentialError a

class Attribute a where

Methods

toAttr :: a -> AttributeValue

Encode an attribute value.

parseAttr :: AttributeValue -> Maybe a

Decode an attribute value.